+++ vortex-array/src/display/extractors/mod.rs
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+mod buffer;
+mod encoding_summary;
+mod metadata;
+mod nbytes;
+mod stats;
+
+pub use buffer::BufferExtractor;
+pub use encoding_summary::EncodingSummaryExtractor;
+pub use metadata::MetadataExtractor;
+pub use nbytes::NbytesExtractor;
+pub use stats::StatsExtractor;
