+++ vortex-array/src/display/extractors/stats.rs
@@ -4,9 +4,6 @@
 use std::fmt::Write;
 use std::fmt::{self};
 
-use humansize::DECIMAL;
-use humansize::format_size;
-
 use crate::DynArray;
 use crate::display::extractor::TreeContext;
 use crate::display::extractor::TreeExtractor;
@@ -114,26 +111,6 @@ impl fmt::Display for StatsDisplay<'_> {
     }
 }
 
-/// Extractor that adds `nbytes=X (Y%)` to the header line.
-pub struct NbytesExtractor;
-
-impl TreeExtractor for NbytesExtractor {
-    fn header_annotations(&self, array: &dyn DynArray, ctx: &TreeContext) -> Vec<String> {
-        let nbytes = array.nbytes();
-        let total_size = ctx.parent_total_size().unwrap_or(nbytes);
-        let percent = if total_size == 0 {
-            0.0
-        } else {
-            100_f64 * nbytes as f64 / total_size as f64
-        };
-        vec![format!(
-            "nbytes={} ({:.2}%)",
-            format_size(nbytes, DECIMAL),
-            percent
-        )]
-    }
-}
-
 /// Extractor that adds stats annotations (e.g. `[nulls=3, min=5]`) to the header line.
 pub struct StatsExtractor;
 
@@ -146,74 +123,5 @@ impl TreeExtractor for StatsExtractor {
         } else {
             vec![trimmed.to_string()]
         }
-    }
-}
-
-/// Extractor that adds a `metadata: ...` detail line.
-pub struct MetadataExtractor;
-
-impl TreeExtractor for MetadataExtractor {
-    fn detail_lines(&self, array: &dyn DynArray, _ctx: &TreeContext) -> Vec<String> {
-        // Capture the metadata_fmt output
-        let mut buf = String::new();
-        // metadata_fmt writes directly to a Formatter, so we use a helper wrapper
-        struct FmtCapture<'a>(&'a dyn DynArray);
-        impl fmt::Display for FmtCapture<'_> {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                self.0.metadata_fmt(f)
-            }
-        }
-        let _ = write!(&mut buf, "{}", FmtCapture(array));
-        vec![format!("metadata: {buf}")]
-    }
-}
-
-/// Extractor that adds buffer detail lines.
-pub struct BufferExtractor {
-    /// Whether to show buffer-level percentage of parent nbytes.
-    pub show_percent: bool,
-}
-
-impl TreeExtractor for BufferExtractor {
-    fn detail_lines(&self, array: &dyn DynArray, _ctx: &TreeContext) -> Vec<String> {
-        let nbytes = array.nbytes();
-        let mut lines = Vec::new();
-        for (name, buffer) in array.named_buffers() {
-            let loc = if buffer.is_on_device() {
-                "device"
-            } else if buffer.is_on_host() {
-                "host"
-            } else {
-                "location-unknown"
-            };
-            let align = if buffer.is_on_host() {
-                buffer.as_host().alignment().to_string()
-            } else {
-                String::new()
-            };
-
-            if self.show_percent {
-                let buffer_percent = if nbytes == 0 {
-                    0.0
-                } else {
-                    100_f64 * buffer.len() as f64 / nbytes as f64
-                };
-                lines.push(format!(
-                    "buffer: {} {loc} {} (align={}) ({:.2}%)",
-                    name,
-                    format_size(buffer.len(), DECIMAL),
-                    align,
-                    buffer_percent,
-                ));
-            } else {
-                lines.push(format!(
-                    "buffer: {} {loc} {} (align={})",
-                    name,
-                    format_size(buffer.len(), DECIMAL),
-                    align,
-                ));
-            }
-        }
-        lines
     }
 }
